remove test notebook
